Extract provider map and JSON serializer to common utils
